added new Features
