+++ src/Components/CardList.js
@@ -11,6 +11,7 @@ const CardList = ({projects}) => {
 
     return (
         <div className='cardListContainer'>
+
             {cardItem}
         </div>
     );
