+++ src/Components/Search.css
@@ -1,11 +1,12 @@
 
-button{
-background-color: #1B8EF2; /* Green */
+.search-buttons button{
+background-color: #1B8EF2;
   border: none;
   color: #f2f2f2;
   padding: 16px 48px;
   text-align: center;
   text-decoration: none;
+  border-radius: 6px;
   font-size: 16px;
   margin:0px 16px 50px 16px !important;
 }
