+++ src/Containers/App.css
@@ -1,7 +1,55 @@
+@import url('https://fonts.googleapis.com/css?family=Montserrat|Ubuntu&display=swap');
+
 div{
   text-align: center;
 }
 
-h1{
+
+
+h1,h2,h3,h4,h5,h6{
   color: #f2f2f2;
+  font-family: 'Montserrat', Helvetica, sans-serif;
+}
+
+header{
+  position: absolute;
+  top: 50%;
+  width: 100%;
+  margin: auto;
+  transform: translateY(-50%);
+}
+
+header h1{
+  font-size: 50px;
+}
+
+header h2{
+  font-size: 26px;
+}
+
+.home-buttons button{
+  background-color: #1B8EF2;
+  border: none;
+  color: #f2f2f2;
+  padding: 16px 48px;
+  text-align: center;
+  text-decoration: none;
+  border-radius: 6px;
+  font-size: 16px;
+  margin:0px 16px 50px 16px !important;
+}
+
+.close{
+  background: transparent;
+  position: absolute;
+  right:0;
+  border: none;
+  font-weight: 900;
+  color: #f2f2f2;
+  font-style: 'Montserrat';
+  padding: 16px 48px;
+  text-align: center;
+  text-decoration: none;
+  border-radius: 6px;
+  font-size: 20px;
 }
